# KAFKA-17495: Factor out assignor benchmark code into utils class