@squah-confluent 2024-09-09
The ServerSideAssignorBenchmarks don't show a consistent change, which is expected.
The homogeneous TargetAssignmentBuilderBenchmark.build benchmarks show a slight unexpected improvement. I'm not sure why.

#### Benchmarks Before
```
Benchmark                                       (assignmentType)  (assignorType)  (isRackAware)  (memberCount)  (partitionsToMemberRatio)  (subscriptionType)  (topicCount)  Mode  Cnt    Score   Error  Units
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         10         HOMOGENEOUS          1000  avgt    2    1.714          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         50         HOMOGENEOUS          1000  avgt    2    7.491          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         10         HOMOGENEOUS          1000  avgt    2    4.309          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         50         HOMOGENEOUS          1000  avgt    2   14.842          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         10         HOMOGENEOUS          1000  avgt    2    0.820          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         50         HOMOGENEOUS          1000  avgt    2    2.417          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         10         HOMOGENEOUS          1000  avgt    2    1.748          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         50         HOMOGENEOUS          1000  avgt    2    7.715          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         10       HETEROGENEOUS          1000  avgt    2  119.580          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         50       HETEROGENEOUS          1000  avgt    2  161.969          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         10       HETEROGENEOUS          1000  avgt    2  247.979          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         50       HETEROGENEOUS          1000  avgt    2  368.867          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         10       HETEROGENEOUS          1000  avgt    2  104.617          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         50       HETEROGENEOUS          1000  avgt    2  145.051          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         10       HETEROGENEOUS          1000  avgt    2  220.763          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         50       HETEROGENEOUS          1000  avgt    2  307.568          ms/op
JMH benchmarks done

Benchmark                                     (memberCount)  (partitionsToMemberRatio)  (topicCount)  Mode  Cnt   Score   Error  Units
TargetAssignmentBuilderBenchmark.build                 5000                         10          1000  avgt    2   1.320          ms/op
TargetAssignmentBuilderBenchmark.build                 5000                         50          1000  avgt    2   3.402          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         10          1000  avgt    2   3.689          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         50          1000  avgt    2  10.657          ms/op
JMH benchmarks done
```

#### Benchmarks After
```
Benchmark                                       (assignmentType)  (assignorType)  (isRackAware)  (memberCount)  (partitionsToMemberRatio)  (subscriptionType)  (topicCount)  Mode  Cnt    Score   Error  Units
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         10         HOMOGENEOUS          1000  avgt    2    1.705          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         50         HOMOGENEOUS          1000  avgt    2    7.165          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         10         HOMOGENEOUS          1000  avgt    2    3.476          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         50         HOMOGENEOUS          1000  avgt    2   14.251          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         10         HOMOGENEOUS          1000  avgt    2    0.766          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         50         HOMOGENEOUS          1000  avgt    2    2.717          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         10         HOMOGENEOUS          1000  avgt    2    1.771          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         50         HOMOGENEOUS          1000  avgt    2    6.761          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         10       HETEROGENEOUS          1000  avgt    2  133.705          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false           5000                         50       HETEROGENEOUS          1000  avgt    2  172.849          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         10       HETEROGENEOUS          1000  avgt    2  267.341          ms/op
ServerSideAssignorBenchmark.doAssignment                    FULL         UNIFORM          false          10000                         50       HETEROGENEOUS          1000  avgt    2  471.977          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         10       HETEROGENEOUS          1000  avgt    2  119.293          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false           5000                         50       HETEROGENEOUS          1000  avgt    2  156.737          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         10       HETEROGENEOUS          1000  avgt    2  242.026          ms/op
ServerSideAssignorBenchmark.doAssignment             INCREMENTAL         UNIFORM          false          10000                         50       HETEROGENEOUS          1000  avgt    2  327.398          ms/op
JMH benchmarks done

Benchmark                                     (memberCount)  (partitionsToMemberRatio)  (subscriptionType)  (topicCount)  Mode  Cnt    Score   Error  Units
TargetAssignmentBuilderBenchmark.build                 5000                         10         HOMOGENEOUS          1000  avgt    2    1.162          ms/op
TargetAssignmentBuilderBenchmark.build                 5000                         50         HOMOGENEOUS          1000  avgt    2    3.153          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         10         HOMOGENEOUS          1000  avgt    2    2.602          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         50         HOMOGENEOUS          1000  avgt    2    7.929          ms/op
TargetAssignmentBuilderBenchmark.build                 5000                         10       HETEROGENEOUS          1000  avgt    2  116.358          ms/op
TargetAssignmentBuilderBenchmark.build                 5000                         50       HETEROGENEOUS          1000  avgt    2  177.115          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         10       HETEROGENEOUS          1000  avgt    2  255.730          ms/op
TargetAssignmentBuilderBenchmark.build                10000                         50       HETEROGENEOUS          1000  avgt    2  341.819          ms/op
JMH benchmarks done
```